delete: 이미지 서비스 합치기
